#6 - 파일정리
main.dart에 있던 PageViewExample을 별도의 파일으로 옮김
View/ 추가
PageViewExample을 SwipePage로 이름 변경, 조금 더 직관적임 (아마도)
